refactor: extract ComposeUpdateChecker from ComposeUpdateService (PR6)
Conservative split of the 1104-line ComposeUpdateService: the read-only
check/lookup concern moves into a new ComposeUpdateChecker collaborator
(IComposeUpdateChecker), leaving the update orchestration (UpdateProject /
UpdateAll + SSE progress) in ComposeUpdateService, which now delegates the four
check/status methods to the checker.

ComposeUpdateChecker owns: CheckProjectUpdates(+internal), CheckAllProjects,
GetGlobalUpdateStatus, ClearCache, FindComposeFilePath, and the compose-image
parsing / exclusion helpers. It takes only the deps those need (image-digest,
caches, project-matching, docker, sse, options). ComposeUpdateService drops the
moved deps plus two that were entirely unused (IComposeDiscoveryService,
IComposeOperationService): 16 ctor deps -> 10.

ComposeUpdateService 1104 -> 680 lines; ComposeUpdateChecker 495.

Tests (per the standing rule):
- ComposeUpdateCheckerTests: cache delegation + compose-file resolution
  (direct-name match, project-matching fallback, not-found).
- ComposeUpdateServiceTests: forwards the four check methods to the checker.
The digest-check and orchestration paths pull in concrete docker services
(live daemon) and are not unit-tested.

Verify: dotnet build -c Release -> 0 warnings/errors; dotnet test -> 329 passed;
app boots, GET /health 200 (full DI graph incl. the new checker resolves).
